+++ CHANGELOG.md
@@ -3,7 +3,9 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
+- [BREAKING] Schema :[function,multimethod,spec.alpha,fspec.alpha]/:source-form collapsed into :source/form, same with :source/str
 - Fix multimethod methods facts to correctly link methods with its multi definition
+
 ## 0.3.3 (2019-12-09)
 
 - Throw when indexing if org.clojure/tools.namespace is detected on the classpath